+++ simtools/exceptions.py
@@ -3,7 +3,7 @@
 
 Exceptions include:
 
-- ParamFileError - raised during processing a parameter file.
+- FileError - raised during processing a file.
 """
 
 
@@ -12,11 +12,12 @@ class SimToolsError(Exception):
     pass
 
 
-class ParamFileError(SimToolsError):
-    """Error during processing a parameter file."""
+class FileError(SimToolsError):
+    """Error during processing a file."""
 
-    def __init__(self, msg="", filename=None, lineno=None, error_msg=None):
-        if not msg and any((filename, lineno, error_msg)):
+    def __init__(self, msg="", filename=None, lineno=None, error_msg=None,
+                 line=None):
+        if not msg and any((filename, lineno, error_msg, line)):
             if filename:
                 msg = "File '{}' is invalid".format(filename)
             else:
@@ -29,10 +30,13 @@ class ParamFileError(SimToolsError):
                 msg += " ({}).".format(error_msg)
             else:
                 msg += "."
-        super(ParamFileError, self).__init__(msg)
+            if line:
+                msg += "\n{}".format(line)
+        super(FileError, self).__init__(msg)
         self.filename = filename
         self.lineno = lineno
         self.error_msg = error_msg
+        self.line = line
 
 
 class SimToolsWarning(Warning):